+++ test/body-parser.js
@@ -3,6 +3,13 @@ var http = require('http')
 var methods = require('methods')
 var request = require('supertest')
 
+var testAsyncHooks = false
+try {
+  var asyncHooks = require('async_hooks')
+  testAsyncHooks = true
+} catch (ignored) {
+}
+
 var bodyParser = require('..')
 
 describe('bodyParser()', function () {
@@ -50,6 +57,31 @@ describe('bodyParser()', function () {
       .send('{"user":"tobi"}')
       .expect(200, '{"user":"tobi"}', done)
   })
+
+  if (testAsyncHooks) {
+    it('should work with async hooks', function (done) {
+      var _bodyParser = bodyParser()
+      var asyncLocalStorage = new asyncHooks.AsyncLocalStorage()
+
+      var server = http.createServer(function (req, res) {
+        const store = {
+          contextMaintained: true
+        }
+        asyncLocalStorage.run(store, function () {
+          _bodyParser(req, res, function (err) {
+            res.statusCode = err ? (err.status || 500) : 200
+            res.end(err ? err.message : JSON.stringify(asyncLocalStorage.getStore()))
+          })
+        })
+      })
+
+      request(server)
+        .post('/')
+        .set('Content-Type', 'application/json')
+        .send('{"user":"tobi"}')
+        .expect(200, '{"contextMaintained":true}', done)
+    })
+  }
 
   describe('http methods', function () {
     before(function () {