+++ HISTORY.md
@@ -8,6 +8,7 @@ unreleased
     - deps: http-errors@1.7.3
   * deps: safe-buffer@5.2.0
   * deps: type-is@~1.6.18
+  * Make it work with Async Hooks
 
 1.19.0 / 2019-04-25
 ===================